+++ src/rcpp_computePointDensity.h
@@ -1,6 +1,6 @@
 #ifndef rcpp_pointDensity_hpp
 #define rcpp_pointDensity_hpp
 
-double rcpp_pointDensity(double x, double theta, double p, double b, int N);
+double rcpp_pointDensity(double x, double theta, double p, double b, double N);
 
 #endif
